Fix bugs about data type issue within MaxMinLPCentralNode.cpp

+++ max_min_lp_simulation/src/MaxMinLPCentralNode.cpp
@@ -514,28 +514,28 @@ bool MaxMinLPCentralNode::initialize(max_min_lp_simulation::MessageRequest::Requ
 							for (vector<max_min_lp_msgs::general_node>::iterator it = temp_gen_r_node.begin(); it != temp_gen_r_node.end(); ++it) {
 								ROS_INFO("        type = %s, id = %d, loc_deg = %d", it->type.c_str(), it->id, it->loc_deg);
 								for (int j = 0; j < it->loc_deg; j++) {
-									ROS_INFO("            loc_id = %d, loc_edge_weight = %f", it->loc_neighbor[j], it->loc_edge_weight[j]);
+									ROS_INFO("            loc_id = %d, loc_edge_weight = %f", it->loc_neighbor[j], (float)it->loc_edge_weight[j]);
 								}
 							}
 							ROS_INFO("    red nodes");
 							for (vector<max_min_lp_msgs::general_node>::iterator it = temp_gen_p_r_node.begin(); it != temp_gen_p_r_node.end(); ++it) {
 								ROS_INFO("        type = %s, id = %d, loc_deg = %d", it->type.c_str(), it->id, it->loc_deg);
 								for (int j = 0; j < it->loc_deg; j++) {
-									ROS_INFO("            loc_id = %d, loc_edge_weight = %f", it->loc_neighbor[j], it->loc_edge_weight[j]);
+									ROS_INFO("            loc_id = %d, loc_edge_weight = %f", it->loc_neighbor[j], (float)it->loc_edge_weight[j]);
 								}
 							}
 							ROS_INFO("    blue nodes");
 							for (vector<max_min_lp_msgs::general_node>::iterator it = temp_gen_p_t_node.begin(); it != temp_gen_p_t_node.end(); ++it) {
 								ROS_INFO("        type = %s, id = %d, loc_deg = %d", it->type.c_str(), it->id, it->loc_deg);
 								for (int j = 0; j < it->loc_deg; j++) {
-									ROS_INFO("            loc_id = %d, loc_edge_weight = %f", it->loc_neighbor[j], it->loc_edge_weight[j]);
+									ROS_INFO("            loc_id = %d, loc_edge_weight = %f", it->loc_neighbor[j], (float)it->loc_edge_weight[j]);
 								}
 							}
 							ROS_INFO("    target nodes");
 							for (vector<max_min_lp_msgs::general_node>::iterator it = temp_gen_t_node.begin(); it != temp_gen_t_node.end(); ++it) {
 								ROS_INFO("        type = %s, id = %d, loc_deg = %d", it->type.c_str(), it->id, it->loc_deg);
 								for (int j = 0; j < it->loc_deg; j++) {
-									ROS_INFO("            loc_id = %d, loc_edge_weight = %f", it->loc_neighbor[j], it->loc_edge_weight[j]);
+									ROS_INFO("            loc_id = %d, loc_edge_weight = %f", it->loc_neighbor[j], (float)it->loc_edge_weight[j]);
 								}
 							}
 						}
@@ -591,7 +591,7 @@ vector<max_min_lp_msgs::general_node> MaxMinLPCentralNode::buildGeneralNode(stri
 			if (m_verbal_flag) {
 				ROS_INFO("        type = %s, id = %d, loc_deg = %d", "robot", m_robot_id[i], (int)m_robots_to_primitives[i].size());
 				for (vector<int>::iterator it = m_robots_to_primitives[i].begin(); it != m_robots_to_primitives[i].end(); ++it) {
-					ROS_INFO("            loc_id = %d, loc_edge_weight = %f", *it, 1);
+					ROS_INFO("            loc_id = %d, loc_edge_weight = %f", *it, (float)1);
 				}
 			}
 		}
@@ -617,7 +617,7 @@ vector<max_min_lp_msgs::general_node> MaxMinLPCentralNode::buildGeneralNode(stri
 			if (m_verbal_flag) {
 				ROS_INFO("        type = %s, id = %d, loc_deg = %d", "red", m_primitive_id[i], (int)m_primitives_to_robots[i].size());
 				for (vector<int>::iterator it = m_primitives_to_robots[i].begin(); it != m_primitives_to_robots[i].end(); ++it) {
-					ROS_INFO("            loc_id = %d, loc_edge_weight = %f", *it, 1);
+					ROS_INFO("            loc_id = %d, loc_edge_weight = %f", *it, (float)1);
 				}
 			}
 		}
